Add Zone.js awareness (#52)
**Why**

Angular has [a concept](https://angular.io/api/core/ApplicationRef#isStable)
of an application being `stable`. This is an important concept of Angular
applications as it is used for testing purposes, the registration of custom
timers and more. Unfortunately `stable` is implemented via
[Zones](https://github.com/angular/angular/blob/master/packages/zone.js).
More specifically an Angular application is defined as stable when there are
no more pending tasks.

Weasel does register some timers, e.g. when making `XMLHttpRequest` or `fetch`
API calls. These timers in turn are blocking the Angular applicatiom from
becoming stable.

**What**

Angular has its own zones that it uses to execute the stabilization checks.
Try to always avoid these zones by entering the root zone when running on
a website that is leveraging the Zone library. This in turn means that
any website employing a similar mechanism will:

 a) Not see our timers.
 b) Potentially run faster because no change identification logic is executed
    due to our timers.