Fix wrong timings
Why
===

We have seen cases in which the reported timings are wrong. Specifically
`appCache` and `SSL` timings.

What
====

`appCache` timings can be wrong when the `fetchStart` is not available.
According to the specification this can happen when the following
condition does not hold true for a navigation…

 > If the new resource is to be fetched using HTTP GET or equivalent […]
 > Source: https://www.w3.org/TR/navigation-timing/#dom-performancetiming-fetchstart

Similarly the SSL timings can be huge in some browser. We are not
entirely sure when this can happen, but are suspecting cases where
an existing TCP connection is re-used and upgraded to a TLS connection
(although this shouldn't happen).