Stop flow workers after running `npm run test:flow` (#87)
The flow command in the test:flow command (implicitly triggerd by
npm test) starts a bunch of flow servers (depending on the number of
available cores) and leaves them running afterwards. This might make
sense if you work on this package directly, to speed up consecutive
builds by a few milliseconds. But weasel is routinely included in other
projects as a sub-dependency, which also means triggering weasel builds
as part of the build of that larger system. Leaving ~20 processes behind
after such a build is not great.

Starting the flow servers every time they are needed (and stopping them
afterwards) does not take long and it seems to be an acceptable
tradeoff.